feat: introduce presentation layer documentation and refactor imports
- Added a comprehensive README.md for the presentation layer, detailing its structure, components, and architecture.
- Refactored import paths in several video-related components and hooks to utilize a centralized utils directory for grid source and detection color utilities.
- Implemented new test files for service context and various hooks to ensure functionality and reliability.
- Enhanced the testing infrastructure with mock implementations for video grid components and toast notifications.
