+++ src/front-end/src/presentation/hooks/useCameraTransport.ts
@@ -2,8 +2,8 @@ import { useCallback, useRef } from 'react';
 import type { Dispatch, RefObject } from 'react';
 import type { ActiveFilterState } from '@/presentation/components/filters/FilterPanel';
 import type { IToastDisplay } from '@/infrastructure/transport/transport-types';
-import type { GridSource, GridSourceAction } from '@/presentation/components/video/grid-source';
-import { GridSourceActionType } from '@/presentation/components/video/grid-source';
+import type { GridSource, GridSourceAction } from '@/presentation/utils/grid-source';
+import { GridSourceActionType } from '@/presentation/utils/grid-source';
 import { webrtcService } from '@/infrastructure/connection/webrtc-service';
 import { logger } from '@/infrastructure/observability/otel-logger';
 import { ChunkReassembler } from '@/infrastructure/transport/data-channel-framing';
