+++ src/front-end/src/presentation/hooks/useSourceFilterSync.ts
@@ -2,9 +2,9 @@ import { useEffect } from 'react';
 import type { Dispatch, RefObject } from 'react';
 import type { ActiveFilterState } from '@/presentation/components/filters/FilterPanel';
 import type { IToastDisplay } from '@/infrastructure/transport/transport-types';
-import type { GridSource, GridSourceAction } from '@/presentation/components/video/grid-source';
-import { GridSourceActionType } from '@/presentation/components/video/grid-source';
-import { normalizeFilters } from '@/presentation/components/video/grid-source';
+import type { GridSource, GridSourceAction } from '@/presentation/utils/grid-source';
+import { GridSourceActionType } from '@/presentation/utils/grid-source';
+import { normalizeFilters } from '@/presentation/utils/grid-source';
 import { webrtcService } from '@/infrastructure/connection/webrtc-service';
 import { logger } from '@/infrastructure/observability/otel-logger';
 import type { useCameraTransport } from '@/presentation/hooks/useCameraTransport';
