+++ src/front-end/src/presentation/test-utils/register-mock-video-grid.ts
