+++ src/front-end/src/presentation/utils/detection-colors.ts
