+++ src/front-end/src/presentation/utils/grid-source.ts
